refactor: prettier format, move variant categories
