+++ src/modules/category/variant/variantCategory.controller.ts
@@ -10,10 +10,10 @@ import {
   Inject,
   NotFoundException,
 } from '@nestjs/common';
-import { VariantCategoryService } from './variantCategory.service';
+import { VariantCategoryService } from '../../variantCategory/variantCategory.service';
 import { CreateVariantCategoryDto } from './dto/create-variantCategory.dto';
 import { UpdateVariantCategoryDto } from './dto/update-variantCategory.dto';
-import { IVariantCategoryService } from './variantCategory.interface';
+import { IVariantCategoryService } from '../../variantCategory/variantCategory.interface';
 import { VariantCategory } from '@prisma/client';
 
 @Controller('/variant-categories')
