+++ src/modules/category/variant/variantCategory.interface.ts
