+++ src/modules/category/variant/variantCategory.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { VariantCategoryService } from './variantCategory.service';
+import { VariantCategoryService } from '../../variantCategory/variantCategory.service';
 import { VariantCategoryController } from './variantCategory.controller';
 import { PrismaModule } from 'src/providers/prisma.module';
