+++ src/modules/product/product.interface.ts
@@ -26,14 +26,17 @@ export interface IProductService {
    * @returns {Promise<Array<Product>>}
    */
   getProducts(): Promise<Array<Product>>;
-  
-    /**
+
+  /**
    * Actualiza un producto por su identificador.
    *
-   * @param {number} id 
-   * @param {UpdateProductDto} updateProductDto 
+   * @param {number} id
+   * @param {UpdateProductDto} updateProductDto
    *
    * @returns {Promise<Product>}
    */
-    updateProductById(id: number, updateProductDto: UpdateProductDto): Promise<Product>;
+  updateProductById(
+    id: number,
+    updateProductDto: UpdateProductDto,
+  ): Promise<Product>;
 }
